refactor: extract workflow run relative code to custom hook
